Updated Javascript further.
This update includes not only side to side arrow key movement of the ball, but two stepping stones for the ball to bounce on.

I hope to, in the future, make the ball stop the animation and return to the play button upon the ball making contact with the floor. with maybe a game over alert message, or something along those lines.